Organized Student Forms and Changed layouts
